Mockchain.hs revamped
